+++ packages/stream_chat/lib/src/core/http/stream_http_client.dart
@@ -3,14 +3,13 @@ import 'dart:async';
 import 'package:dio/dio.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
-import 'package:stream_core/stream_core.dart' show ApiErrorInterceptor, DioExceptionMapping, SystemEnvironmentManager;
+import 'package:stream_core/stream_core.dart'
+    show ApiErrorInterceptor, AuthInterceptor, DioExceptionMapping, SystemEnvironmentManager, TokenManager;
 import '../error/error.dart';
 import 'connection_id_manager.dart';
 import 'interceptor/additional_headers_interceptor.dart';
-import 'interceptor/auth_interceptor.dart';
 import 'interceptor/connection_id_interceptor.dart';
 import 'interceptor/logging_interceptor.dart';
-import 'token_manager.dart';
 
 part 'stream_http_client_options.dart';
 
@@ -45,7 +44,7 @@ class StreamHttpClient {
       }
       ..interceptors.addAll([
         AdditionalHeadersInterceptor(systemEnvironmentManager),
-        if (tokenManager != null) AuthInterceptor(this, tokenManager),
+        if (tokenManager != null) AuthInterceptor(httpClient, tokenManager),
         if (connectionIdManager != null) ConnectionIdInterceptor(connectionIdManager),
         const ApiErrorInterceptor(),
         ...interceptors ??